move docs test
