+++ go.mod
@@ -2,7 +2,4 @@ module github.com/cszczepaniak/go-sqlbuilder
 
 go 1.20
 
-require (
-	github.com/google/go-cmp v0.6.0
-	github.com/mattn/go-sqlite3 v1.14.17
-)
+require github.com/google/go-cmp v0.6.0
