+++ integration/docs_test.go
@@ -1,4 +1,4 @@
-package sqlbuilder
+package integration
 
 import (
 	"bufio"
