+++ integration/docs_test.tmpl
@@ -1,4 +1,4 @@
-package sqlbuilder
+package integration
 
 {{ .Imports }}
